scramble() neu geschrieben Noch einen Sicherheitscheck in die NMR-Routine Öfter mal log_error statt printf benutzen

+++ src/common/kernel/resources.c
@@ -63,8 +63,8 @@ terraform_resources(region * r)
 			res->type       = tdata->rawmaterials[i].type;
 			res->amount     = (int)(res->base * (1+(res->level-res->startlevel)*(res->divisor/100.0)));
 #ifdef RESOURCE_FACTOR
-			res->amount     = res->amount * RESOURCE_FACTOR;
-			if(res->amount == 0) res->amount = 1;
+			res->amount     = (int)(res->amount * RESOURCE_FACTOR);
+			if (res->amount == 0) res->amount = 1;
 #endif
 			assert(res->amount>0);
 			res->type->terraform(res, r);
